% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{add_args_to_dataframe}
\alias{add_args_to_dataframe}
\title{Adds arguments supplied explicity to dataframe.}
\usage{
add_args_to_dataframe(x, ..., replace = TRUE)
}
\arguments{
\item{x}{Info dataframe.}

\item{...}{Arguments that are added as columns to dataframe.}

\item{replace}{Replace existing columns? If set to \code{FALSE}
default arguments can be supplied.}
}
\description{
Most functions have default values whoch can be overwritten.
}
\keyword{internal}

