% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{add_proj_data}
\alias{add_proj_data}
\title{Add projections as user coords based on spherical coords}
\usage{
add_proj_data(s, proj = "", parameters = NULL, orientation = NULL)
}
\description{
Uses last projection \code{.Last.projection()} to calc projected
coordinates. Adds the variables \code{xp} and \code{yp} to dataframe
based on spherical coordinates \code{phi} and \code{theta}.
}
\keyword{internal}

