% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{add_sphere_coords}
\alias{add_sphere_coords}
\title{Add colummns with spherical coordinates}
\usage{
add_sphere_coords(x, r = 1)
}
\description{
Add colummns with spherical coordinates
}
\keyword{internal}

