% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{cart_to_surface_of_sphere}
\alias{cart_to_surface_of_sphere}
\title{3d cartesians to sperical with fixed radius}
\usage{
cart_to_surface_of_sphere(x, r = 1)
}
\description{
3d cartesians to sperical with fixed radius
}
\keyword{internal}

