% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{cbind_replace}
\alias{cbind_replace}
\title{#Draw a circle
#
#Wrapper for \code{\link{plotrix::draw.circle}}
#
Column binding of new columns, replacement of existing ones}
\usage{
cbind_replace(x, y)
}
\description{
#Draw a circle
#
#Wrapper for \code{\link{plotrix::draw.circle}}
#
Column binding of new columns, replacement of existing ones
}
\examples{
a <- data.frame(a="a1")
ab <- data.frame(a="a2", b="b1")
c <- data.frame(c="c1")
bc <- data.frame(b="b2", c="c2")

cbind_replace(a, c)
cbind_replace(a, ab)
cbind_replace(ab, bc)
}
\keyword{internal}

