% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{color_qual}
\alias{color_qual}
\title{Qualitative color function}
\usage{
color_qual(x)
}
\arguments{
\item{x}{An integer.}
}
\description{
RColorBrewer only has color palettes with a limited
amount of qualitative colors per palette. Here, the
qualitative palettes are concatenated to get more colors.
Of course, there will be more similar colors.
}
\examples{
n <- 60
pie(rep(1,n), col=color_qual(1:60))
}

