% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{c_cart}
\alias{c_cart}
\alias{c_proj}
\alias{c_sphere}
\title{Get coordinate columns from dataframe}
\usage{
c_cart(x)

c_sphere(x)

c_proj(x)
}
\arguments{
\item{x}{A dataframe.}
}
\description{
Comvencience wrapper to get cartesian (\code{}, \code{}, \code{}),
spherical (\code{phi}, \code{theta}, \code{r}), or
projected (\code{xp}, \code{yp}) coordinate columns.
}

