% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gridmap-package.R
\docType{package}
\name{gridmap}
\alias{gridmap}
\alias{gridmap-package}
\title{\pkg{gridmap} draws several graphical elements in polar fashion}
\description{
\pkg{gridmap} draws several graphical elements in polar fashion
}
\note{
Some functions to map vectors and maps.
}
\author{
Mark Heckmann
}
\keyword{gridmap}
\keyword{package}

