% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{interpolate_polygon_points}
\alias{interpolate_polygon_points}
\title{Interpolate points that form a polygon}
\usage{
interpolate_polygon_points(x, n = 100)
}
\arguments{
\item{x}{A dataframe with xyz colummns (must have three columns).}

\item{n}{Number of points to interpolate.}
}
\description{
Interpolate points that form a polygon
}
\keyword{internal}

