% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{map_labels}
\alias{map_labels}
\alias{map_text}
\title{#Draw points on map
#
Add point labels to map}
\usage{
map_labels(x, subset = .(), ...)

map_text(x, ..., subset = .())
}
\arguments{
\item{x}{A dataframe.}

\item{subset}{Logical condition to select subset. Either as a string or using
the \code{\link{plyr::.}} function.}

\item{...}{Arguments that are added as columns to dataframe.}
}
\description{
\code{map_labels} positions labels using an pabels positioning algorithm.
\code{map_text} also plots text but passes all arguments directly to \code{text} while
 in \code{map_labels} they are mapped from the dataframe.
}

