% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{map_centroid_star}
\alias{map_centroid_star}
\title{Draw a star (lines from centroid to points)}
\usage{
map_centroid_star(x, subset = .(), ...)
}
\description{
Draw a star (lines from centroid to points)
}

