% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{map_convex_hull}
\alias{map_convex_hull}
\title{Map convex hull on globe}
\usage{
map_convex_hull(x, subset = .(), n = 100, ...)
}
\arguments{
\item{x}{A dataframe containing xyz and coordinates}

\item{subset}{}

\item{n}{Number of points used to draw each polygon line.}

\item{...}{Arguments passed to \code{polygon}.}

\item{subset}{Logical condition to select subset. Either as a string or using
the \code{\link{plyr::.}} function.}
}
\description{
Map convex hull on globe
}

