% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{map_data}
\alias{map_data}
\title{Prepares the data for plotting}
\usage{
map_data(x, preset = NULL, proj = "", parameters = NULL,
  orientation = NULL)
}
\description{
The function needs a dataframe as input. The arguments that are accepted
cartesian coordinates \code{x}, \code{y}, \code{z},) spherical coordinates
(\code{phi}, \code{theta}, \code{r}), or projected user coordinates
(\code{xp}, \code{yp}). If all are supplied, they are respected in reverse order,
i.e. projected user coordinates are used if given.
}

