% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{map_lines}
\alias{map_lines}
\title{Draw lines between all points in x and y}
\usage{
map_lines(x, y, n = 200, ...)
}
\arguments{
\item{x,y}{Either a matrix ot vector with three columns (x,y,z).
Spherical and projected coords are currently not supported).}

\item{n}{Number of segments used to draw the lines.}
}
\description{
Draw lines between all points in x and y
}

