% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{map_proj}
\alias{map_proj}
\title{Set projection and orientation}
\usage{
map_proj(preset = NULL, proj = "", parameters = NULL,
  orientation = NULL)
}
\description{
Wrapper for \code{mapproj::mapproject}
}

