% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{map_range}
\alias{map_range}
\title{Get limits in user coords for current projection}
\usage{
map_range(by = 1)
}
\arguments{
\item{by}{Sequence step width for lattitude and longitude.}
}
\value{
A list with the elements \code{xlim} and \code{ylim}.
}
\description{
To determine \code{xlim} and \code{ylim} for a current projection
a brute force approach is used. All combinations of a sequence of
lattitude and longitude coordinates is projected and the x and y limits
are returned.
}
\examples{
map_proj("mollweide")
map_range()
}

