% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapster-package.R
\docType{package}
\name{mapster}
\alias{mapster}
\alias{mapster-package}
\title{\pkg{mapster} draws several graphical elements in polar fashion}
\description{
\pkg{mapster} draws several graphical elements in polar fashion
}
\note{
Some functions to map vectors and maps.
}
\author{
Mark Heckmann
}
\keyword{mapster}
\keyword{package}

