% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subset.R
\name{stop_if_not_all_subset_vars_in_df}
\alias{stop_if_not_all_subset_vars_in_df}
\title{Stop if not all variables in logical expression exist in dataframe}
\usage{
stop_if_not_all_subset_vars_in_df(x, e)
}
\arguments{
\item{x}{A dataframe.}

\item{e}{Logical expression as a string or wrappd in \code{plyr::.} function.}
}
\description{
Function checks if all variables used in logical subset expression are
contained in dataframe. If not an error is thrown and the missing variables
are listed.
}
\examples{
\dontrun{
stop_if_not_all_subset_vars_in_df(mtcars, .(mpg > 20))
stop_if_not_all_subset_vars_in_df(mtcars, .(mpg2 > 20))
}
}
\keyword{internal}

