% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subset.R
\name{subset2}
\alias{subset2}
\title{Subsetting with some special features}
\usage{
subset2(x, subset = .(), i = NULL)
}
\arguments{
\item{x}{A dataframe.}

\item{subset}{Logical condition to select subset. Either as a string or using
the \code{\link{plyr::.}} function.}
}
\description{
Subsetting with some special features
}
\examples{
subset2(mtcars, .(mpg > 20))
subset2(mtcars, "mpg > 20 & gear == 4")
}

