% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_indexes}
\alias{subset_indexes}
\title{working horse for \code{subset_indexes}}
\usage{
subset_indexes(x, subset = .())
}
\arguments{
\item{x}{A dataframe.}

\item{subset}{Logical condition to select subset. Either as a string or using
the \code{\link{plyr::.}} function.}
}
\description{
Get dataframe indexed matching logical expression.
}
\keyword{internal}

