% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subset.R
\name{vars_in_expression}
\alias{vars_in_expression}
\title{Extract variable names used in logical expression}
\usage{
vars_in_expression(e = .())
}
\arguments{
\item{e}{Logical expression as a string or wrappd in \code{plyr::.} function.}
}
\value{
A character vector containing the variable names.
}
\description{
The left part of possibly concatenated logical expressions are extracted. Use
is intended for variable names in dataframes. Alphanumerics, dots and
underscores are allowed as variable names.
}
\examples{
vars_in_expression(.(a.1==1 & b==2))
}
\keyword{internal}

