% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{vec_to_xyz_df}
\alias{vec_to_xyz_df}
\title{Convert a vector of length three into dataframe with columns x, y, z.}
\usage{
vec_to_xyz_df(x)
}
\description{
Convert a vector of length three into dataframe with columns x, y, z.
}
\keyword{internal}

