% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{warn_if_var_is_factor}
\alias{warn_if_var_is_factor}
\title{issue warning if specified variables in dataframe are factors}
\usage{
warn_if_var_is_factor(x, vars = NULL)
}
\description{
issue warning if specified variables in dataframe are factors
}
\keyword{internal}

