% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subset.R
\name{which2}
\alias{which2}
\title{\code{which} with different treatment of NULL}
\usage{
which2(x)
}
\arguments{
\item{x}{A logical vector.}
}
\description{
\code{which} function that returns \code{integer(0)} when \code{x=NULL}.
}
\examples{
which2(c(T,F,T))
which2(NULL)
}
\keyword{internal}

